requirements.txt
absl-py==2.3.1
academy-py==0.3.0
accelerate==1.9.0
accumulation_tree==0.6.4
acres==0.5.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiologic==0.14.0
aiosignal==1.4.0
alembic==1.16.4
anndata==0.12.1
annotated-types==0.7.0
anyio==4.9.0
array-api-compat==1.12.0
attrs==25.3.0
beautifulsoup4==4.13.4
biomart==0.9.2
biomcp-python==0.4.6
bitsandbytes==0.42.0
black==25.1.0
certifi==2025.7.14
cffi==2.0.0
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.2.1
cloudpickle==3.1.1
colorlog==6.9.0
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.2
crc32c==2.7.1
cryptography==46.0.1
culsans==0.9.0
cycler==0.12.1
datasets==4.0.0
decorator==5.2.1
dill==0.3.8
diskcache==5.6.3
donfig==0.8.1.post1
duckduckgo_search==8.1.1
etelemetry==0.3.1
filelock==3.18.0
flake8==7.3.0
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.3.0
geneformer @ file:///Users/tnandi/Downloads/agents/agentic_lab/Geneformer
globus-sdk==3.63.0
graphviz==0.21
grpcio==1.74.0
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.5
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.3
idna==3.10
importlib_resources==6.5.2
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
legacy-api-wrap==1.4.1
llvmlite==0.44.0
loompy==3.0.8
looseversion==1.3.0
lxml==6.0.0
Mako==1.3.10
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mccabe==0.7.0
mcp==1.12.2
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
multiprocess==0.70.16
mypy_extensions==1.1.0
natsort==8.4.0
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
numba==0.61.2
numcodecs==0.16.1
numpy==2.3.3
numpy-groupies==0.11.3
optuna==4.4.0
optuna-integration==4.4.0
packaging==25.0
pandas==2.3.2
pathlib==1.0.1
pathspec==0.12.1
patsy==1.0.1
peft==0.16.0
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
primp==0.15.0
propcache==0.3.2
protobuf==6.31.1
prov==2.1.1
proxystore==0.8.3
psutil==7.0.0
puremagic==1.30
pyarrow==21.0.0
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydot==4.0.1
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pynndescent==0.5.13
pyparsing==3.2.3
PyPDF2==3.0.1
pytest==8.4.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pyudorandom==1.0.0
pyxnat==1.6.3
PyYAML==6.0.2
ray==2.48.0
rdflib==7.1.4
redis==6.4.0
referencing==0.36.2
regex==2025.7.31
requests==2.32.4
rich==14.1.0
rpds-py==0.26.0
safetensors==0.5.3
scanpy==1.11.3
scikit-learn==1.7.1
scipy==1.15.3
scprep==1.2.3
seaborn==0.13.2
session-info2==0.2
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.42
sse-starlette==3.0.2
starlette==0.47.2
statsmodels==0.14.5
sympy==1.14.0
tdigest==0.5.2.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tokenizers==0.21.4
tomli_w==1.2.0
torch==2.7.1
tqdm==4.67.1
traits==7.0.2
transformers==4.54.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
umap-learn==0.5.9.post2
urllib3==2.5.0
uvicorn==0.35.0
Werkzeug==3.1.3
wrapt==1.17.3
xxhash==3.5.0
yarl==1.20.1
zarr==3.1.0
